+++ src/components/products/ProductForm.css
@@ -1,0 +1,3 @@
+.product-form-input {
+    width: 100%;
+}
