+++ src/components/products/ProductList.css
@@ -1,0 +1,3 @@
+.empty-icon {
+    font-size: 64px;
+}
